pipeline.py
import torch
import cv2
import numpy as np
from PIL import Image
import config
import models
import utils
import generator
import style_transfer
from yolov5 import YOLOv5  # yolov5 라이브러리 가정


def main():
    # 1. 모델 로드
    print("Loading models...")
    # torch.load 보안 패치 (노트북 내용 반영)
    torch.load = utils._load_patched if hasattr(utils, "_load_patched") else torch.load

    yolo_model = YOLOv5(config.YOLO_MODEL_PATH)

    cls_model = models.ClassificationModel(num_classes=9).to(config.DEVICE)
    cls_model.load_state_dict(
        torch.load(config.CLS_MODEL_PATH, map_location=config.DEVICE)
    )
    cls_model.eval()

    reg_model = models.CoordinateModel(num_classes=9, num_coords=8).to(config.DEVICE)
    reg_model.load_state_dict(
        torch.load(config.REG_MODEL_PATH, map_location=config.DEVICE)
    )
    reg_model.eval()

    vgg = models.get_vgg_encoder().to(config.DEVICE)
    decoder = models.get_decoder().to(config.DEVICE)
    # ... (VGG/Decoder weight 로드 로직 추가) ...
    vgg.eval()
    decoder.eval()

    # 2. 이미지 로드 및 객체 탐지
    img_path = "test_image.jpg"  # 테스트 이미지
    original_image = Image.open(img_path)

    results = yolo_model.predict(original_image)
    if results.xyxy[0].size(0) == 0:
        print("No license plate detected.")
        return

    bbox = results.xyxy[0][0][:4].detach().cpu().numpy().astype(int)

    # 3. 전처리 및 분류/회귀
    cropped_img, _ = utils.crop_with_margin(original_image, bbox)
    input_tensor = utils.preprocess_image(cropped_img)

    # ... (분류 및 좌표 예측 로직 수행) ...

    # 4. 가상 번호판 생성
    gen = generator.ImageGenerator(config.RESULT_DIR, config.PLATES_DIR_A, ...)
    # fake_plate = gen.generate_type_a() ...

    # 5. 스타일 트랜스퍼 및 합성
    # content = ...
    # style = ...
    # styled_plate = style_transfer.run_style_transfer(...)

    # 6. 결과 저장 및 시각화
    print("Processing complete.")


if __name__ == "__main__":
    main()
