requirements.txt
torch>=1.10.0
torchvision>=0.11.0
opencv-python>=4.5.0
Pillow>=9.0.0
numpy>=1.21.0
timm>=0.6.0           
ultralytics>=8.0.0    
tqdm>=4.64.0          
matplotlib>=3.5.0    
pyyaml>=6.0           
scipy>=1.7.0         